+++ src/containerapp/azext_containerapp/_help.py
@@ -608,6 +608,184 @@ helps['containerapp env workload-profile update'] = """
           az containerapp env workload-profile update -g MyResourceGroup -n MyEnvironment --workload-profile-name my-wlp --workload-profile-type D4 --min-nodes 1 --max-nodes 3
 """
 
+# Container Apps Job Commands
+helps['containerapp job'] = """
+    type: group
+    short-summary: Commands to manage Container Apps jobs.
+"""
+
+helps['containerapp job create'] = """
+    type: command
+    short-summary: Create a container apps job.
+    examples:
+    - name: Create a container apps job with Trigger Type as Manual.
+      text: |
+          az containerapp job create -n MyContainerappsjob -g MyResourceGroup \\
+              --trigger-type Manual \\
+              --replica-timeout 5 \\
+              --replica-retry-limit 2 \\
+              --replica-completion-count 1 \\
+              --parallelism 1 \\
+              --image imageName \\
+              --workload-profile-name my-wlp
+    - name: Create a container apps job with Trigger Type as Schedule.
+      text: |
+          az containerapp job create -n MyContainerappsjob -g MyResourceGroup \\
+              --trigger-type Schedule \\
+              --replica-timeout 5 \\
+              --replica-retry-limit 2 \\
+              --replica-completion-count 1 \\
+              --parallelism 1 \\
+              --cron-expression \"*/1 * * * *\" \\
+              --image imageName
+    - name: Create a container apps job with Trigger Type as Event.
+      text: |
+          az containerapp job create -n MyContainerappsjob -g MyResourceGroup \\
+              --trigger-type Event \\
+              --replica-timeout 5 \\
+              --replica-retry-limit 2 \\
+              --replica-completion-count 1 \\
+              --parallelism 1 \\
+              --polling-interval 30 \\
+              --min-executions 0 \\
+              --max-executions 1 \\
+              --scale-rule-name queueJob \\
+              --scale-rule-type azure-queue \\
+              --scale-rule-metadata "accountName=mystorageaccountname" \\
+                                    "cloud=AzurePublicCloud" \\
+                                    "queueLength": "5" "queueName": "foo" \\
+              --scale-rule-auth "connection=my-connection-string-secret-name" \\
+              --image imageName
+"""
+
+helps['containerapp job update'] = """
+    type: command
+    short-summary: Update a Container Apps Job.
+    examples:
+    - name: Update a job's replica timeout.
+      text: |
+          az containerapp job update -n MyContainerAppJob -g MyResourceGroup \\
+              --replica-timeout 10
+"""
+
+helps['containerapp job delete'] = """
+    type: command
+    short-summary: Delete a Container Apps Job.
+    examples:
+    - name: Delete a job.
+      text: az containerapp job delete -n MyContainerAppJob -g MyResourceGroup
+"""
+
+helps['containerapp job show'] = """
+    type: command
+    short-summary: Show details of a Container Apps Job.
+    examples:
+    - name: Show the details of a job.
+      text: |
+          az containerapp job show -n MyContainerAppJob -g MyResourceGroup
+"""
+
+helps['containerapp job list'] = """
+    type: command
+    short-summary: List Container Apps Job by subscription or resource group.
+    examples:
+    - name: List jobs in the current subscription.
+      text: |
+          az containerapp job list
+    - name: List environments by resource group.
+      text: |
+          az containerapp job list -g MyResourceGroup
+"""
+
+helps['containerapp job start'] = """
+    type: command
+    short-summary: Start a Container Apps Job execution.
+    examples:
+    - name: Start a job execution.
+      text: az containerapp job start -n MyContainerAppJob -g MyResourceGroup
+    - name: Start a job with different image and configurations.
+      text: az containerapp job start -n MyContainerAppJob -g MyResourceGroup --image MyImageName --cpu 0.5 --memory 1.0Gi
+"""
+
+helps['containerapp job stop'] = """
+    type: command
+    short-summary: Stops a Container Apps Job execution.
+    examples:
+    - name: Stop a job execution.
+      text: az containerapp job stop -n MyContainerAppJob -g MyResourceGroup
+    - name: Stop a job execution giving a specific job execution name.
+      text: az containerapp job stop -n MyContainerAppJob -g MyResourceGroup --job-execution-name MyContainerAppJob-66v9xh0
+    - name: Stop multiple job executions giving a list of execution names.
+      text: az containerapp job stop -n MyContainerAppJob -g MyResourceGroup --execution-name-list MyContainerAppJob-66v9xh0,MyContainerAppJob-66v9xh1
+"""
+
+# Container App Job Secret Commands
+helps['containerapp job secret'] = """
+    type: group
+    short-summary: Commands to manage secrets.
+"""
+
+helps['containerapp job secret show'] = """
+    type: command
+    short-summary: Show details of a secret.
+    examples:
+    - name: Show the details of a secret.
+      text: |
+          az containerapp job secret show -n MyContainerappjob -g MyResourceGroup --secret-name MySecret
+"""
+
+helps['containerapp job secret list'] = """
+    type: command
+    short-summary: List the secrets of a container app job.
+    examples:
+    - name: List the secrets of a container app job.
+      text: |
+          az containerapp job secret list -n MyContainerappjob -g MyResourceGroup
+"""
+
+helps['containerapp job secret remove'] = """
+    type: command
+    short-summary: Remove secrets from a container app job.
+    examples:
+    - name: Remove secrets from a container app job.
+      text: |
+          az containerapp job secret remove -n MyContainerappjob -g MyResourceGroup --secret-names MySecret MySecret2
+"""
+
+helps['containerapp job secret set'] = """
+    type: command
+    short-summary: Create/update secrets.
+    examples:
+    - name: Add secrets to a container app job.
+      text: |
+          az containerapp job secret set -n MyContainerappjob -g MyResourceGroup --secrets MySecretName1=MySecretValue1 MySecretName2=keyvaultref:https://example.vault.azure.net/secrets/mysecret,identityref:/subscriptions/sub/resourceGroups/rg/providers/Microsoft.ManagedIdentity/userAssignedIdentities/myidentity
+    - name: Update a secret.
+      text: |
+          az containerapp job secret set -n MyContainerappjob -g MyResourceGroup --secrets MyExistingSecretName=MyNewSecretValue MyExistingSecretName2=keyvaultref:https://example.vault.azure.net/secrets/mysecret,identityref:/subscriptions/sub/resourceGroups/rg/providers/Microsoft.ManagedIdentity/userAssignedIdentities/myidentity
+"""
+
+# Container App Job execution commands
+helps['containerapp job execution'] = """
+    type: group
+    short-summary: Commands to view executions of a Container App Job.
+"""
+
+helps['containerapp job execution list'] = """
+    type: command
+    short-summary: Get list of all executions of a Container App Job.
+    examples:
+    - name: List of all executions of a Container App Job.
+      text: az containerapp job execution list -n MyContainerAppJob -g MyResourceGroup
+"""
+
+helps['containerapp job execution show'] = """
+    type: command
+    short-summary: Get execution of a Container App Job.
+    examples:
+    - name: Get execution of a Container App Job.
+      text: az containerapp job execution show -n MyContainerAppJob -g MyResourceGroup --job-execution-name MyContainerAppJob-66v9xh0
+"""
+
 # Certificates Commands
 helps['containerapp env certificate'] = """
     type: group
@@ -727,6 +905,60 @@ helps['containerapp identity show'] = """
     - name: Show managed identities.
       text: |
           az containerapp identity show -n myContainerapp -g MyResourceGroup
+"""
+
+helps['containerapp job identity'] = """
+    type: group
+    short-summary: Commands to manage managed identities for container app job.
+"""
+
+helps['containerapp job identity assign'] = """
+    type: command
+    short-summary: Assign managed identity to a container app job.
+    long-summary: Managed identities can be user assigned or system assigned.
+    examples:
+    - name: Assign system identity.
+      text: |
+          az containerapp job identity assign -n myContainerappjob -g MyResourceGroup --system-assigned
+    - name: Assign user identity.
+      text: |
+          az containerapp job identity assign -n myContainerappjob -g MyResourceGroup --user-assigned myUserIdentityName
+    - name: Assign user identity (from a different resource group than the containerapp job).
+      text: |
+          az containerapp job identity assign -n myContainerappjob -g MyResourceGroup --user-assigned myUserIdentityResourceId
+    - name: Assign system and user identity.
+      text: |
+          az containerapp job identity assign -n myContainerappjob -g MyResourceGroup --system-assigned --user-assigned myUserIdentityResourceId
+"""
+
+helps['containerapp job identity remove'] = """
+    type: command
+    short-summary: Remove a managed identity from a container app job.
+    examples:
+    - name: Remove system identity.
+      text: |
+          az containerapp job identity remove -n myContainerappjob -g MyResourceGroup --system-assigned
+    - name: Remove user identity.
+      text: |
+          az containerapp job identity remove -n myContainerappjob -g MyResourceGroup --user-assigned myUserIdentityName
+    - name: Remove system and user identity (from a different resource group than the containerapp job).
+      text: |
+          az containerapp job identity remove -n myContainerappjob -g MyResourceGroup --system-assigned --user-assigned myUserIdentityResourceId
+    - name: Remove all user identities.
+      text: |
+          az containerapp job identity remove -n myContainerappjob -g MyResourceGroup --user-assigned
+    - name: Remove system identity and all user identities.
+      text: |
+          az containerapp job identity remove -n myContainerappjob -g MyResourceGroup --system-assigned --user-assigned
+"""
+
+helps['containerapp job identity show'] = """
+    type: command
+    short-summary: Show managed identities of a container app job.
+    examples:
+    - name: Show managed identities.
+      text: |
+          az containerapp job identity show -n myContainerappjob -g MyResourceGroup
 """
 
 # Ingress Commands
@@ -1328,4 +1560,64 @@ helps['containerapp compose create'] = """
           az containerapp compose create -g MyResourceGroup \\
               --environment MyContainerappEnv \\
               --compose-file-path "path/to/docker-compose.yml"
+"""
+
+# Patch commands
+helps['containerapp patch'] = """
+    type: group
+    short-summary: Patch Azure Container Apps. Patching is only available for the apps built using the source to cloud feature. See https://aka.ms/aca-local-source-to-cloud
+"""
+
+helps['containerapp patch list'] = """
+   type: command
+   short-summary: List container apps that can be patched. Patching is only available for the apps built using the source to cloud feature. See https://aka.ms/aca-local-source-to-cloud
+   examples:
+    - name: List patchable container apps in the current subscription.
+      text: |
+          az containerapp patch list
+    - name: List patchable container apps by resource group.
+      text: |
+          az containerapp patch list -g MyResourceGroup
+    - name: List patchable container apps by managed environment.
+      text: |
+          az containerapp patch list -g MyResourceGroup --environment MyContainerAppEnv
+    - name: List patchable and unpatchable container apps by managed environment with the show-all option.
+      text: |
+          az containerapp patch list -g MyResourceGroup --environment MyContainerAppEnv --show-all
+"""
+
+helps['containerapp patch apply'] = """
+    type: command
+    short-summary: List and apply container apps to be patched. Patching is only available for the apps built using the source to cloud feature. See https://aka.ms/aca-local-source-to-cloud
+    examples:
+    - name: List patchable container apps in the current subscription and apply patch.
+      text: |
+          az containerapp patch apply
+    - name: List patchable container apps by resource group and apply patch.
+      text: |
+          az containerapp patch apply -g MyResourceGroup
+    - name: List patchable container apps by managed environment and apply patch.
+      text: |
+          az containerapp patch apply -g MyResourceGroup --environment MyContainerAppEnv
+    - name: List patchable and unpatchable container apps by managed environment with the show-all option and apply patch for patchable container apps.
+      text: |
+          az containerapp patch apply -g MyResourceGroup --environment MyContainerAppEnv --show-all
+"""
+
+helps['containerapp patch interactive'] = """
+    type: command
+    short-summary: List and select container apps to be patched in an interactive way. Patching is only available for the apps built using the source to cloud feature. See https://aka.ms/aca-local-source-to-cloud
+    examples:
+    - name: List patchable container apps in the current subscription and apply patch interactively.
+      text: |
+          az containerapp patch interactive
+    - name: List patchable container apps by resource group and apply patch interactively.
+      text: |
+          az containerapp patch interactive -g MyResourceGroup
+    - name: List patchable container apps by managed environment and apply patch interactively.
+      text: |
+          az containerapp patch interactive -g MyResourceGroup --environment MyContainerAppEnv
+    - name: List patchable and unpatchable container apps by managed environment with the show-all option and apply patch for patchable container apps interactively.
+      text: |
+          az containerapp patch interactive -g MyResourceGroup --environment MyContainerAppEnv --show-all
 """