+++ src/containerapp/azext_containerapp/_models.py
@@ -138,6 +138,13 @@ ServiceBinding = {
     "name": None
 }
 
+JobScale = {
+    "minExecutions": None,
+    "maxExecutions": None,
+    "pollingInterval": None,
+    "rules": []  # list of ScaleRule
+}
+
 TrafficWeight = {
     "revisionName": None,
     "weight": None,
@@ -189,6 +196,49 @@ Configuration = {
     "registries": None  # [RegistryCredentials]
 }
 
+JobTemplate = {
+    "containers": None,  # [Container]
+    "initContainers": None,  # [Container]
+    "volumes": None  # [Volume]
+}
+
+# Added template for starting job executions
+JobExecutionTemplate = {
+    "template": {
+        "containers": None,  # [Container]
+        "initContainers": None  # [Container]
+    }
+}
+
+JobConfiguration = {
+    "secrets": None,  # [Secret]
+    "triggerType": None,  # 'manual' or 'schedule' or 'event'
+    "replicaTimeout": None,
+    "replicaRetryLimit": None,
+    "manualTriggerConfig": None,  # ManualTriggerConfig
+    "scheduleTriggerConfig": None,  # ScheduleTriggerConfig
+    "eventTriggerConfig": None,  # EventTriggerConfig
+    "registries": None,  # [RegistryCredentials]
+    "dapr": None
+}
+
+ManualTriggerConfig = {
+    "replicaCompletionCount": None,
+    "parallelism": None
+}
+
+ScheduleTriggerConfig = {
+    "replicaCompletionCount": None,
+    "parallelism": None,
+    "cronExpression": None
+}
+
+EventTriggerConfig = {
+    "replicaCompletionCount": None,
+    "parallelism": None,
+    "scale": None,  # [JobScale]
+}
+
 UserAssignedIdentity = {
 
 }
@@ -222,6 +272,18 @@ ContainerApp = {
         "environmentId": None,
         "configuration": None,  # Configuration
         "template": None,  # Template
+        "workloadProfileName": None
+    },
+    "tags": None
+}
+
+ContainerAppsJob = {
+    "location": None,
+    "identity": None,  # ManagedServiceIdentity
+    "properties": {
+        "environmentId": None,
+        "configuration": None,  # JobConfiguration
+        "template": None,  # JobTemplate
         "workloadProfileName": None
     },
     "tags": None
@@ -315,4 +377,32 @@ ManagedCertificateEnvelop = {
         "subjectName": None,  # str
         "validationMethod": None  # str
     }
+}
+
+# ContainerApp Patch
+ImageProperties = {
+    "imageName": None,
+    "targetContainerName": None,
+    "targetContainerAppName": None,
+    "revisionMode": None,
+}
+
+ImagePatchableCheck = {
+    "targetContainerAppName": None,
+    "targetContainerName": None,
+    "revisionMode": None,
+    "targetImageName": None,
+    "oldRunImage": None,
+    "newRunImage": None,
+    "id": None,
+    "reason": None,
+}
+
+OryxMarinerRunImgTagProperty = {
+    "fullTag": None,
+    "framework": None,
+    "version": None,
+    "marinerVersion": None,
+    "architectures": None,
+    "support": None,
 }