+++ src/containerapp/HISTORY.rst
@@ -2,8 +2,29 @@
 
 Release History
 ===============
-Upcoming
+0.3.33
 ++++++
+* 'az containerapp create': fix --registry-identity "system" with --revision-suffix
+* 'az containerapp up': fix --target-port value not being propagated when buildpack is used to build image from --source
+* Fix for 'az containerapp job create' with --yaml option to create a Container App job
+* Support 'az containerapp job secret' to manage secrets for Container App jobs
+* Support 'az containerapp job identity' to manage identity for Container App jobs
+* Fix for issue with --user-assigned identity for Container App jobs where identities were getting split incorrectly
+* Add new parameters `--mi-system-assigned` and `--mi-user-assigned` to replace the deprecated parameters `--system-assigned` and `--user-assigned` for `az containerapp job create` command
+
+0.3.32
+++++++
+* Fix for 'az containerapp job update' command when updating Container App job with a trigger configuration
+
+0.3.31
+++++++
+* Fix issue when using 'az containerapp up' to create a container app from a local source with a Dockerfile
+
+0.3.30
+++++++
+* Add 'az containerapp service' for binding a service to a container app
+* Add 'az containerapp patch' to enable the local source to cloud
+* Add 'az containerapp job' to manage Container Apps jobs
 * Split 'az containerapp env workload-profile set' into 'az containerapp env workload-profile add' and 'az containerapp env workload-profile update'
 * Add 'az containerapp env workload-profile add' to support creating a workload profile in an environment
 * Add 'az containerapp env workload-profile update' to support updating an existing workload profile in an environment